requirements.txt
autopep8==2.3.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
Deprecated==1.2.14
feedparser==6.0.11
flake8==7.1.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-logging==3.11.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.9
importlib_metadata==8.4.0
mccabe==0.7.0
opentelemetry-api==1.27.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
sgmllib3k==1.0.0
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
